Added getComputerChoice function
Began project with function mentioned above that will serve the purpose of randomly giving us the computers choice in rock paper scissors.

Declared variable computerChoice that had math floor and math random stored in to return a random interger. Then used if statements which would print out the choice of the computer which would be random each time.