Play a single round of rock, paper, scissors
Added playRound function with two parameters for the computer and the player. Promt used to ask the user for their choice and then used if and && statements to compare the two results and then return the result of the round with a string such as 'you lose'. Input from the player selection also made case insensitive.