Set up java file to give rock, paper scissors a UI
Removed 5 round logic & created three buttons for each selection. Added an even listener to the buttons that called the playRound function each time a selection was made. Added a results div and then converted the console.logs into DOM methods that displayed the number of rounds, the outcome of the round, the score and the outcome of the final result once someone reached 5.